updating api

+++ MacroMateApp/ViewModels/DailyLogViewModel.cs
@@ -52,10 +52,10 @@ namespace MacroMateApp.ViewModels
 
         // Temporary properties for user input
         public string FoodName { get; set; } = "";
-        public int Calories { get; set; }
-        public int Protein { get; set; }
-        public int Carbs { get; set; }
-        public int Fats { get; set; }
+        public double Calories { get; set; }
+        public double Protein { get; set; }
+        public double Carbs { get; set; }
+        public double Fats { get; set; }
         public string SelectedMeal { get; set; } = "Breakfast"; // Default selection
